fixed type errors

+++ app/(kambaz)/courses/[cid]/modules/page.tsx
@@ -43,11 +43,15 @@ export default function Modules() {
   }; 
 
   const onRemoveModule = async (moduleId: string) => { 
+    if (!cid) return;
+
     await client.deleteModule(cid, moduleId); 
     dispatch(setModules(modules.filter((m: any) => m._id !== moduleId))); 
   }; 
 
   const onUpdateModule = async (module: any) => { 
+    if (!cid) return;
+    
     await client.updateModule(cid, module); 
     const newModules = modules.map((m: any) => m._id === module._id ? module : m ); 
     dispatch(setModules(newModules));
